Basic interoperability between Java8 Streams and Scala collections.
This consists of
  (1) toScala[Coll] extension methods for Java8 Streams
  (2) seqStream and parStream extension methods for Scala collections
  (3) A manually specialized set of Accumulators that let you quickly
      save a stream for multiple re-use.
    (a) accumulate and accumulatePrimitive methods on Java8 Streams
    (b) to[Coll] on Accumulators
    (c) .iterator, .toArray, .toList also

There is a lot of redundant code in the Accumulators since it is difficult
to pull out common functionality without endangering performance of the
manually specialized cases.

Everything goes through the new Stepper trait to partially tame the
Spliterator hierarchy.

At this point, Scala collections all go through Accumulator.  Need to write
something custom for the more important collections.

Tests written.

Scaladocs written for the new classes.